test/requirements.txt
websocket-client
asyncio

